Changes after the peer review
New changes made:
1-The three option buttons were formatted to look more visible.
	Their font was increased, space added, and a hoover action was added (it shows the silhouette).

2-“+ Report” button.
This button used to be a link to the Report page, the action was changed to increase by one the Number of Reports. It can be clicked one time only, after that the button is disabled.

3-Fix the FAQs page scroll bar.

4-Add day and month for First Occurrence. Ex: 05/09/2022
The date field in the initial database was changed to the format mm/dd/yyyy. When a new scam email is added to the list of emails the date will be formatted in the same way.

5-The table should have no sorting at the beginning, only when the user clicks on a header, it should sort.
	Originally, the first time the home page was opened it would show the list of scam emails sorted by Number of Reports. Now, the list is presented unsorted.

6-Add option to sort by Email and Type of Scam
		The list of scam emails can now be sorted by Email and Type of Scam.